requirements.txt
streamlit
pandas
yahooquery
openai
python-dotenv